Add all-profiles flag to collect all profiles at once
For those familiar, this emulates the behavior of older tool
`gimme-aws-creds`.

When this flag is given, it will get temporary credentials for all
accounts and roles assigned to the user, and save them to a profile
named "{idp}-{role}". If the account has an alias (and the role has
permissions to fetch its own account alias), then the profile instead
will be named "{alias}-{role}".

This flag implies "aws-credentials" format as "env-var" makes no sense
for more than a single profile.